Changed timer command to dailyprompt and set timezone to EST
Also added an arbitrary prefix that must be set equal to 'set' because the client requested it. I could not really think of a better way to do this, and using message.startswith() is even worse as a practice, so this is the best way to meet the client's requests.