Added optional role parameter and active_prompt functionality.
Added the option to pass in a role that will be mentioned. Also, added the functionality to tell if there is an active prompt. In the event that there is an active prompt, the command will yield an error to the Discord channel it was called from and the timer will not be set until it the current task is completed, or the current task is disabled and a new one is created.